pyproject.toml
[project]
name = "hcbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-genai>=1.21.1",
    "streamlit>=1.46.0",
]
